Add capture of run_id metadata
Most (but not all) benchmark files have a five character (almost) unique ID autogenerated when the benchmarks are run: "echoResults2019-11-27_YidUg.txt". Trevor added a function to pull out this (almost) unique id as a piece of metadata called "run_id". This will allow all the results from a given run to be presented together.