Add parsing of compiler string to capture the generator and platform metadata
For now, since the platform metadata can be missing (empty string), I'm matching the generator data on a pre-defined list and finding the platform by process of elimination (may be empty string). I log an error if no match on the generator is made so we know to add it to the pre-defined list as necessary.