Add "unspecified" core_type to all benchmarks when none is specified
This allows the construction of datasets for analysis where all benchmarks should have a core_type that this scripts has provided and allows those values to be filtered out more easily when slicing-n-dicing datasets to graph. If a core_type of "nan" is showing up in the results, it means something has gone wrong somewhere.