Add filter to not create graphs when the core_type is "unspecified"
There are some test results that have the "unspecified" core type and since I was iterating over all possible core types (regardless of the benchmark) when deciding which graphs to create I ended up creating graphs that had null data. Now, by redefining the dataframe to only be the benchmark of interest only the core types for that benchmark are considered.